+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Other
 
+- [#950](https://github.com/ClementTsang/bottom/pull/950): Update help menu for disk and temperature widgets to support sorting.
+
 ## [0.7.1] - 2023-01-06
 
 ## Bug Fixes